Moved all splineboris-adjacent files to beam_elements/splineboris_src
